+++ classes/Parser.php
@@ -52,6 +52,8 @@ class Parser implements ParserInterface
     {
         $this->config = $config;
         $this->transport = $transport;
+        // @deprecated 2.0.0
+        $this->props = [];
     }
 
     /**
@@ -63,7 +65,7 @@ class Parser implements ParserInterface
      *
      * @return array Processed content and shortcodes
      */
-    public function interpretShortcodes(string $content, string $id, array $page)
+    public function processShortcodes(string $content, string $id, array $page)
     {
         $handlers = new HandlerContainer();
         $path = $page['path'];
@@ -221,5 +223,112 @@ class Parser implements ParserInterface
         $unwrap = self::REGEX_MEDIA_P;
         $content = preg_replace($unwrap, "$1", $content);
         return $content;
+    }
+
+    /**
+     * Parse shortcodes
+     *
+     * @param string $content Markdown content in Page
+     * @param string $id      Slide id-attribute
+     *
+     * @deprecated 2.0.0
+     *
+     * @return array Processed contents and properties
+     */
+    public function interpretShortcodes(string $content, string $id)
+    {
+        $handlers = new HandlerContainer();
+        $handlers->setDefault(
+            function (ShortcodeInterface $sc) {
+                $return = array();
+                $name = $sc->getName();
+                $value = $sc->getParameter($name, $sc->getBbCode());
+                if (Utils::startsWith($name, 'class')) {
+                    $this->props['class'] = $value;
+                } elseif (Utils::startsWith($name, 'style')) {
+                    $name = str_replace('style-', '', $name);
+                    $this->props['styles'][$name] = $value;
+                } elseif (Utils::startsWith($name, 'data')) {
+                    $this->props['styles'][$name] = $value;
+                } elseif (Utils::startsWith($name, 'hide')) {
+                    $this->props['hide'] = [true];
+                }
+                return;
+            }
+        );
+        $parser = "Thunder\Shortcode\Parser\\" . $this->config['shortcode_parser'];
+        $processor = new Processor(new $parser(), $handlers);
+        return ['content' => $processor->process($content), 'props' => $this->props];
+    }
+    
+    /**
+     * Create HTML for fragments
+     *
+     * @param string $content Markdown content in Page
+     *
+     * @deprecated 2.0.0
+     *
+     * @return string Processed contents
+     */
+    public function processFragments(string $content)
+    {
+        $content = preg_replace(
+            self::REGEX_FRAGMENT_SHORTCODE,
+            '<span class="fragment \\1">\\2</span>',
+            $content
+        );
+        return $content;
+    }
+
+    /**
+     * Process styles and data-attributes
+     *
+     * @param array  $styles List of key-value pairs
+     * @param string $route  Route to Page for relative assets
+     * @param string $id     Slide id-attribute
+     * @param string $base   Base path to prepend to file
+     *
+     * @deprecated 2.0.0
+     *
+     * @return string Processed styles, in inline string
+     */
+    public function processStylesData(array $styles, string $route, string $id, string $base = "")
+    {
+        $inline = $data = '';
+        foreach ($styles as $property => $value) {
+            if ($property == 'background-image') {
+                if (!Uri::isValidUrl($value)) {
+                    $locations = array(
+                        '',
+                        'user/pages',
+                        'user/pages/images',
+                    );
+                    $locations = Utilities::explodeFileLocations($locations, GRAV_ROOT, '/', '/');
+                    $file = Utilities::fileFinder($value, $locations);
+                    $file = str_ireplace(GRAV_ROOT, '', $file);
+                    $value = $base . $file;
+                }
+                $inline .= $property . ': url(' . $value . ');';
+            } elseif (Utils::startsWith($property, 'data')) {
+                $data .= ' ' . $property . '="' . $value . '"';
+                if ($property == 'data-textsize-scale') {
+                    $this->transport->setClass($id, 'textsizing');
+                }
+            } elseif ($property == 'header-font-family') {
+                $this->transport->setStyle($id, "{\nfont-family:$value;\n}", 'h1,h2,h3,h4,h5,h6');
+            } elseif ($property == 'header-color') {
+                $this->transport->setStyle($id, "{\ncolor:$value;\n}", 'h1,h2,h3,h4,h5,h6');
+            } elseif ($property == 'block-font-family') {
+                $this->transport->setStyle($id, "{\nfont-family:$value;\n}");
+            } elseif ($property == 'block-color') {
+                $this->transport->setStyle($id, "{\ncolor:$value;\n}");
+            } else {
+                $inline .= $property . ': ' . $value . ';';
+            }
+        }
+        return array(
+            'style' => $inline,
+            'data' => $data
+        );
     }
 }